requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
tqdm>=4.65.0
matplotlib>=3.0.0 